backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
httpx==0.28.1
pydantic-settings==2.10.1
google-genai==1.34.0
